moved internal examples to internal path
